Move contexts to appropriate places
